feat: convert bhop to ESP for CS2
- I replaced the CS:GO bunnyhop functionality in the usermode application with a console-based ESP for CS:2.
- I added new header files for CS2 offsets and class definitions based on your provided dumper output.
- I included a vector math header for World-to-Screen calculations.
- I updated the driver communication wrapper to be 64-bit compatible.
- The new ESP reads the entity list, filters for enemies, and prints their screen coordinates and health to the console.

+++ usermode example/driver.hpp
@@ -9,7 +9,7 @@ class driver_manager {
 	HANDLE m_driver_handle = nullptr; //handle to our driver
 
     struct info_t { //message type that will be passed between user program and driver
-        UINT64 target_pid = 0; //process id of process we want to read from / write to
+        uintptr_t target_pid = 0; //process id of process we want to read from / write to
         UINT64 target_address = 0x0; //address in the target proces we want to read from / write to
         UINT64 buffer_address = 0x0; //address in our usermode process to copy to (read mode) / read from (write mode)
         UINT64 size = 0; //size of memory to copy between our usermode process and target process
@@ -17,12 +17,12 @@ class driver_manager {
     };
 
 public:
-    driver_manager(const char* driver_name, DWORD target_process_id) {
+    driver_manager(const char* driver_name, uintptr_t target_process_id) {
 		m_driver_handle = CreateFileA(driver_name, GENERIC_READ, 0, nullptr, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, nullptr); //get a handle to our driver
         attach_to_process(target_process_id); //tell driver to find the process with this id and put it into g_target_process
 	}
 
-    void attach_to_process(DWORD process_id) {
+    void attach_to_process(uintptr_t process_id) {
         info_t io_info;
         
         io_info.target_pid = process_id;
